[Client]: Moved client predicted data in client_predicted_state_t struct, being used as 'predictedStates[CMD_BACKUP]' and distinct 'predictedState' [SVGames]: The stored delta_yaw which was an integer, now is a float. (The int was leftover from short coordinates).

+++ src/baseq2/svgame/g_phys.cpp
@@ -367,7 +367,7 @@ typedef struct {
     vec3_t  origin;
     vec3_t  angles;
 #if USE_SMOOTH_DELTA_ANGLES
-    int     deltayaw;
+    float deltayaw;
 #endif
 } pushed_t;
 pushed_t    pushed[MAX_EDICTS], *pushed_p;
